wm: More work on panel docks.
This is far from done, but I figure I should send it
out before it gets larger.  This adds a visible background
to panel docks, resizes toplevel windows when docks are
shown, and adds rudimentary support for packing all of the
panels in a dock together.  There are still a bunch of
things missing, like the ability to reorder panels within
a dock, some sort of visible separation between the panels,
etc.

Review URL: http://codereview.chromium.org/602025