assets: Update Chromium OS boot image.
This is just a simple blue starburst image that I got
from Cole.  We might want to replace it with an image
that's closer to Chrome's login background and that
says "Chromium OS" at some point, but it's an
improvement over the bright-blue image that we have
right now.

BUG=chromium-os:3224
TEST=tried it

Review URL: http://codereview.chromium.org/2811027