assets: Add "Chrome Droid Sans" font, a Droid Sans variant.
This has a simplified version of the 'g' character.  We may
want to look at trying to upstream this into the main Droid
Sans font at some point, but it seems easiest to just
release our own version for now.

I'll update the chromeos-asset package's ebuilds in a
later change so that these files are included.

BUG=chromium-os:5210
TEST=none

Review URL: http://codereview.chromium.org/2883050