assets: improve startup message appearance
The boot splash has changed to white background, so we should change the output
format (black on white) to fit the new theme.  Border and padding are no longer
required.

Render engine is also changed.  Pango renders some fonts (especially for RTL)
better so we use it by default now.

BUG=chromium-os:20671,chrome-os-partner:3052
TEST=emerge-x86-generic chromeos-assets

Change-Id: Ieb0f9903e2d41f68f9b3b30b051e36de4affedfb
Reviewed-on: http://gerrit.chromium.org/gerrit/8193
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>