Update ChromeVox from upstream (milestone 18 version)
The major new feature in this version of ChromeVox is that it uses its own indicator as it moves throughout a document, it no longer uses WebKit selection. This is a visual improvement, plus it improves compatibility with some pages that disabled selection or managed selection for some other reason.

In addition, ChromeVox supports a number of new Chrome OS private APIs, to give feedback on changes to the volume, brightness, screen lock, and update status.

There are many other small features and bug fixes.

BUG=chromium-os:25472
TEST=Manual testing by pressing Ctrl+Alt+Z on login screen to enable spoken feedback

Change-Id: I9d26d4a9c0f5a031764280f6e258ee347af458be
Reviewed-on: https://gerrit.chromium.org/gerrit/14907
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Zachary Kuznia <zork@chromium.org>
Tested-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>