Provide a C.UTF-8 locale
Loaded by packaging with the extension and backed by URLLoader in the
NaCl module.

Locale files themselves have to be checked in because building them is
tricky. The NaCl SDK provides the input files, but no compiled locale
files. Furthermore, the locale files are not stable across glibc
versions, so we specificly need to build with eglibc 2.9's localedef.
Unfortunately, this too is not provided by the SDK. See
http://code.google.com/p/nativeclient/issues/detail?id=2542

Locale files were built by checking out eglibc 2.9 sources, building
localedef with --prefix=$(NACL_TOOLCHAIN_ROOT) and running

  localedef -c -f UTF-8 -i POSIX ./C.UTF-8

We may be able to simplify things by only using one file,
/lib/locale/locale-archive. But glibc only knows how to read that with
mmap, so we'd have to emulate enough of that.